Refactor code

+++ src/main/java/simulator/entity/Grid.java
@@ -2,8 +2,9 @@ package simulator.entity;
 
 import javax.swing.*;
 import java.awt.*;
+import java.awt.geom.Line2D;
+import java.awt.geom.Rectangle2D;
 import java.util.ArrayList;
-import java.awt.geom.*;
 
 public class Grid extends JPanel {
 
@@ -43,26 +44,26 @@ public class Grid extends JPanel {
             for (Cell cell : row) {
                 if (cell.isObstaclePresent()) {
                     switch (cell.getImageDirection()) {
-                        case NONE:
-                            break;
-                        case NORTH:
-                            obstacleFront.add(new MyPoint(cell.getX() + size.width / 2,
-                                    cell.getY() + size.height / 2 - offset, cell.getImageDirection()));
-                            break;
-                        case SOUTH:
-                            obstacleFront.add(new MyPoint(cell.getX() + size.width / 2,
-                                    cell.getY() + size.height / 2 + offset, cell.getImageDirection()));
-                            break;
-                        case EAST:
-                            obstacleFront.add(new MyPoint(cell.getX() + size.width / 2 + offset,
-                                    cell.getY() + size.height / 2, cell.getImageDirection()));
-                            break;
-                        case WEST:
-                            obstacleFront.add(new MyPoint(cell.getX() + size.width / 2 - offset,
-                                    cell.getY() + size.height / 2, cell.getImageDirection()));
-                            break;
-                        default:
-                            break;
+                    case NONE:
+                        break;
+                    case NORTH:
+                        obstacleFront.add(new MyPoint(cell.getX() + size.width / 2,
+                                cell.getY() + size.height / 2 - offset, cell.getImageDirection()));
+                        break;
+                    case SOUTH:
+                        obstacleFront.add(new MyPoint(cell.getX() + size.width / 2,
+                                cell.getY() + size.height / 2 + offset, cell.getImageDirection()));
+                        break;
+                    case EAST:
+                        obstacleFront.add(new MyPoint(cell.getX() + size.width / 2 + offset,
+                                cell.getY() + size.height / 2, cell.getImageDirection()));
+                        break;
+                    case WEST:
+                        obstacleFront.add(new MyPoint(cell.getX() + size.width / 2 - offset,
+                                cell.getY() + size.height / 2, cell.getImageDirection()));
+                        break;
+                    default:
+                        break;
 
                     }
 
@@ -144,22 +145,22 @@ public class Grid extends JPanel {
     public boolean checkIfNeedReverse(MyPoint p, int turningRadius) {
         p = (MyPoint) p.clone();
         switch (p.getDirection()) {
-            case NONE:
-                break;
-            case NORTH:
-                p.translate(0, (int) -turningRadius);
-                return this.checkIfPointCollides(p);
-            case SOUTH:
-                p.translate(0, (int) turningRadius);
-                return this.checkIfPointCollides(p);
-            case EAST:
-                p.translate((int) turningRadius, 0);
-                return this.checkIfPointCollides(p);
-            case WEST:
-                p.translate((int) -turningRadius, 0);
-                return this.checkIfPointCollides(p);
-            default:
-                break;
+        case NONE:
+            break;
+        case NORTH:
+            p.translate(0, -turningRadius);
+            return this.checkIfPointCollides(p);
+        case SOUTH:
+            p.translate(0, turningRadius);
+            return this.checkIfPointCollides(p);
+        case EAST:
+            p.translate(turningRadius, 0);
+            return this.checkIfPointCollides(p);
+        case WEST:
+            p.translate(-turningRadius, 0);
+            return this.checkIfPointCollides(p);
+        default:
+            break;
 
         }
